+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Current release candidate
 * Added support for LightGBM models
 * Added support for XGBoost models
 * Added support for PyTorch models
+* Added support for CatBoost models
 
 0.2.1 (2019-11-19)
 ------------------